process.py
# Questo è solo uno pseudocodice e un esempio concettuale

from PIL import Image
import pytesseract # Per OCR
from transformers import pipeline # Per NLP (Hugging Face)
from deep_translator import GoogleTranslator

 # Per traduzione (o DeepL, o Hugging Face)


pytesseract.pytesseract.tesseract_cmd = r'/opt/homebrew/Cellar/tesseract/5.5.0_1/bin/tesseract' # Assicurati di avere Tesseract installato

def process_document(file_path, target_language='it'):
    text = ""
    # 1. OCR (se necessario)
    if file_path.lower().endswith(('.png', '.jpg', '.jpeg', '.tiff', '.bmp')):
        image = Image.open(file_path)
        text = pytesseract.image_to_string(image, lang='ita') # 'ita' per italiano
    elif file_path.lower().endswith(('.pdf')):
        # Per PDF, avresti bisogno di librerie come PyPDF2 o pdfminer.six
        # e poi fare OCR pagina per pagina se il PDF non è testuale
        # Oppure usare un servizio cloud OCR che gestisce i PDF
        print("Il supporto PDF richiede librerie aggiuntive o servizi cloud.")
        return "Errore: Supporto PDF non implementato in questo esempio."
    else: # Assumiamo sia un file di testo puro
        with open(file_path, 'r', encoding='utf-8') as f:
            text = f.read()

    if not text:
        return "Nessun testo estratto dal documento."

    # 2. Analisi NLP per estrazione e riassunto
    # Per una soluzione reale, qui useresti un modello fine-tuned
    # In questo esempio, usiamo un modello generico di riassunto
    summarizer = pipeline("summarization", model="facebook/bart-large-cnn") # Modello generico per riassunto abstractive
    # Oppure per estrazione di informazioni (NER o QA)
    # nlp_model = pipeline("ner", model="dbmdz/bert-base-italian-xxl-uncased") # Esempio per italiano NER

    summary_result = summarizer(text, max_length=150, min_length=30, do_sample=False)
    extracted_summary = summary_result[0]['summary_text']

    # Per un modello più specifico, potresti voler fare:
    # 1. Identificazione del tipo di documento
    # 2. Estrazione di entità specifiche (date, nomi, importi, azioni richieste)
    # 3. Costruzione di un riassunto strutturato basato sulle entità estratte

    final_summary = []
    # Esempio di come potresti strutturare l'output se estraessi punti specifici
    # (Questo richiederebbe un modello NLP molto più specifico addestrato sui tuoi dati)
    # if "contratto" in text.lower(): # Semplice classificazione per parola chiave
    #     final_summary.append("Tipo di documento: Contratto di Lavoro")
    #     final_summary.append("Punti Chiave:")
    #     final_summary.append("- Data di inizio: [DATA_ESTRATTA]")
    #     final_summary.append("- Mansioni: [MANSIONI_ESTRATTE]")
    #     final_summary.append("- Documenti da firmare: [DOCUMENTI_ESTRATTI]")
    #     # ... e così via
    # else:
    final_summary.append("Riassunto Generale:")
    final_summary.append(extracted_summary)


    # 3. Traduzione (se la lingua target è diversa dall'italiano)
    if target_language != 'it':
        translator = GoogleTranslator(source='auto', target=target_language)
        translated_summary_parts = []
        for part in final_summary:
            if "Riassunto Generale:" in part:
                translated_summary_parts.append(part)
            else:
                translated_summary_parts.append(translator.translate(part))
        final_summary = translated_summary_parts


    return "\n".join(final_summary)

# Esempio di utilizzo:
# Assumi di avere un file 'documento_isee.txt' o 'contratto.png'
# summary = process_document('documento_esempio.txt', target_language='en')
# print(summary)

# summary_spanish = process_document('documento_esempio.txt', target_language='es')
# print(summary_spanish)

#summary = process_document("test.jpeg", target_language='en')
#print(summary)