requirements.txt
flask
pytesseract
pdf2image
transformers
torch
Pillow
deep-translator
